requirements.txt
kiteconnect>=5.0.1
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.26.0
APScheduler>=3.10.0
requests>=2.31.0
